backend/requirements.txt
Django==5.0.9
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
